Trade accounts controller refactored to include calculated balance

+++ API/Entities/Account.cs
@@ -7,8 +7,8 @@ namespace API.Entities
         public string Name { get; set; }
         public string Nickname { get; set; }
         public DateTime CreatedDate { get; set; } = DateTime.UtcNow;
-        public int InitialBalance { get; set; }
-        public int Balance { get; set; }
+        public float InitialBalance { get; set; }
+        public float Balance { get; set; }
         public List<Trade> Trades { get; set; } = [];
     }
 }
